9-binary_tree_height.c: data type Edited

+++ 9-binary_tree_height.c
@@ -7,8 +7,8 @@
 
 size_t binary_tree_height(const binary_tree_t *tree)
 {
-	int leftHeight = 0;
-	int rightHeight = 0;
+	size_t leftHeight = 0;
+	size_t rightHeight = 0;
 
 	if (tree == NULL)
 		return (0);
